# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BuildCv.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5062/"
    }
  ],
  "paths": {
...
    "/v1/resumes/{id}": {
      "get": {
        "tags": [
          "Resumes"
        ],
        "summary": "Returns one CV in full, every entry carrying the id that addresses it.",
        "description": "THE ONLY ROUTE THAT CARRIES ENTRY IDS. Each item of each collection has an `id` that is stable for as long as that entry exists, and it is what `DELETE /v1/resumes/{id}/{section}/{itemId}` takes. Ids are unique within one CV and opaque otherwise: they are not dense, not ordered, and an entry deleted and re-added gets a new one. Do not address an entry by its position in these arrays — the store returns each collection as a set, so a position can name a different entry between two reads. `GET /v1/resumes` deliberately does NOT carry the collections or their ids; fetch this route for the CV a candidate is about to edit.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resumes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AwardResponse": {
        "required": [
          "id",
          "title",
          "awarder",
          "date",
          "summary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "awarder": {
            "type": [
              "null",
              "string"
            ]
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CertificateResponse": {
        "required": [
          "id",
          "name",
          "issuer",
          "credentialId",
          "credentialUrl",
          "validityPeriod"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "credentialId": {
            "type": [
              "null",
              "string"
            ]
          },
          "credentialUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "validityPeriod": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DateRangeResponse"
              }
            ]
          }
        }
      },
...
      "ContactInformationResponse": {
        "required": [
          "fullName",
          "email",
          "phoneNumber",
          "location",
          "website",
          "summary",
          "profiles"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "website": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "profiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileResponse"
            }
          }
        }
      },
...
      "DateRangeResponse": {
        "required": [
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "start": {
            "type": "string"
          },
          "end": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EducationResponse": {
        "required": [
          "id",
          "institution",
          "degree",
          "fieldOfStudy",
          "period",
          "grade",
          "level"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "institution": {
            "type": "string"
          },
          "degree": {
            "type": [
              "null",
              "string"
            ]
          },
          "fieldOfStudy": {
            "type": [
              "null",
              "string"
            ]
          },
          "period": {
            "$ref": "#/components/schemas/DateRangeResponse"
          },
          "grade": {
            "type": [
              "null",
              "string"
            ]
          },
          "level": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ExperienceResponse": {
        "required": [
          "id",
          "type",
          "organization",
          "position",
          "period",
          "summary",
          "highlights"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/DateRangeResponse"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "InterestResponse": {
        "required": [
          "id",
          "name",
          "keywords"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "LanguageResponse": {
        "required": [
          "id",
          "name",
          "fluency",
          "level"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "fluency": {
            "type": [
              "null",
              "string"
            ]
          },
          "level": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProfileResponse": {
        "required": [
          "network",
          "username",
          "url"
        ],
        "type": "object",
        "properties": {
          "network": {
            "type": "string"
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProjectResponse": {
        "required": [
          "id",
          "name",
          "period",
          "description",
          "repositoryUrl",
          "liveDemoUrl",
          "technologies",
          "highlights"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/DateRangeResponse"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "repositoryUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "liveDemoUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "technologies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PublicationResponse": {
        "required": [
          "id",
          "title",
          "publisher",
          "url",
          "releaseDate",
          "summary"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "publisher": {
            "type": [
              "null",
              "string"
            ]
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReferenceResponse": {
        "required": [
          "id",
          "name",
          "position",
          "company",
          "email",
          "phoneNumber",
          "referenceText"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "position": {
            "type": [
              "null",
              "string"
            ]
          },
          "company": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phoneNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "referenceText": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ResumeResponse": {
        "required": [
          "id",
          "ownerId",
          "name",
          "contactInformation",
          "createdAt",
          "updatedAt",
          "experiences",
          "educations",
          "skills",
          "projects",
          "certificates",
          "languages",
          "awards",
          "publications",
          "interests",
          "references"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "contactInformation": {
            "$ref": "#/components/schemas/ContactInformationResponse"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "experiences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceResponse"
            }
          },
          "educations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EducationResponse"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponse"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            }
          },
          "certificates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificateResponse"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LanguageResponse"
            }
          },
          "awards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AwardResponse"
            }
          },
          "publications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationResponse"
            }
          },
          "interests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterestResponse"
            }
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenceResponse"
            }
          }
        }
      },
...
      "SkillResponse": {
        "required": [
          "id",
          "name",
          "level",
          "yearsOfExperience",
          "keywords"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": [
              "null",
              "string"
            ]
          },
          "yearsOfExperience": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
...
}